apply change code review

+++ app/src/main/java/fr/coppernic/samples/core/ui/ApiMappingFragment.kt
@@ -1,12 +1,15 @@
 package fr.coppernic.samples.core.ui
 
 import android.content.Intent
+import android.nfc.Tag
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Button
+import android.widget.ImageButton
 import fr.coppernic.samples.core.R
 import fr.coppernic.samples.core.ui.key.KeyContent
 import fr.coppernic.samples.core.ui.key.KeyItem
@@ -15,6 +18,7 @@ import fr.coppernic.samples.core.ui.screen.ShortcutActivity
 import fr.coppernic.sdk.mapping.Mapper
 import fr.coppernic.sdk.mapping.cone2.MapperImpl
 import kotlinx.android.synthetic.main.fragment_api_mapping.*
+import java.util.function.Consumer
 
 /**
  * A simple [Fragment] subclass.
@@ -23,11 +27,8 @@ class ApiMappingFragment : Fragment() {
 
     private val TAG = "LoadProgKey"
 
-    lateinit var mapper: Mapper
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
+    lateinit var mapper: Mapper
 
     override fun onResume() {
         super.onResume()
@@ -42,7 +43,6 @@ class ApiMappingFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        loadMapper()
     }
 
     private fun loadMapper() {
@@ -50,22 +50,18 @@ class ApiMappingFragment : Fragment() {
             mapper = it
             refresh()
         }, {
-            throw it
+            Log.e(TAG, it.message)
         })
     }
 
     private fun loadFromMapper(progKey: Mapper.ProgKey) {
         val type = mapper.getMappingType(progKey)
-        lateinit var actionName: String
+        var actionName = "";
 
         when (type) {
             Mapper.MappingType.KEY -> actionName = loadKey(mapper.getKeyMapping(progKey))
             Mapper.MappingType.SHORTCUT -> actionName = mapper.getShortcutMapping(progKey)
             else -> Log.d(TAG, "Not supported yet")
-        }
-
-        if (actionName == null) {
-            actionName = ""
         }
 
         when (progKey) {
@@ -82,52 +78,45 @@ class ApiMappingFragment : Fragment() {
     }
 
     private fun loadScreen() {
-
-        btnKey1.setOnClickListener {
-            val intent = Intent(context, KeyActivity::class.java)
-            intent.putExtra(KEY, P1)
-            startActivity(intent)
-        }
-        btnShortcut1.setOnClickListener {
-            val intent = Intent(context, ShortcutActivity::class.java)
-            intent.putExtra(SHORTCUT, P1)
-            startActivity(intent)
-        }
-        img_btn_deleteP1.setOnClickListener {
-            mapper.remove(Mapper.ProgKey.P1)
-            refresh()
+        var keyPos = 1;
+        listOf<Button>(btnKey1, btnKey2, btnKey3).forEach {
+            val pNameKey = "P${keyPos++}"
+            it.setOnClickListener {
+                val intent = Intent(context, KeyActivity::class.java)
+                intent.putExtra(KEY, pNameKey)
+                startActivity(intent)
+            }
         }
 
-
-        btnKey2.setOnClickListener {
-            val intent = Intent(context, KeyActivity::class.java)
-            intent.putExtra(KEY, P2)
-            startActivity(intent)
-        }
-        btnShortcut2.setOnClickListener {
-            val intent = Intent(context, ShortcutActivity::class.java)
-            intent.putExtra(SHORTCUT, P2)
-            startActivity(intent)
-        }
-        img_btn_deleteP2.setOnClickListener {
-            mapper.remove(Mapper.ProgKey.P2)
-            refresh()
+        var shortPos = 1
+        listOf<Button>(btnShortcut1, btnShortcut2, btnShortcut3).forEach {
+            val pNameShort = "P${shortPos++}"
+            it.setOnClickListener {
+                val intent = Intent(context, ShortcutActivity::class.java)
+                intent.putExtra(SHORTCUT, pNameShort)
+                startActivity(intent)
+            }
         }
 
-
-        btnKey3.setOnClickListener {
-            val intent = Intent(context, KeyActivity::class.java)
-            intent.putExtra(KEY, P3)
-            startActivity(intent)
-        }
-        btnShortcut3.setOnClickListener {
-            val intent = Intent(context, ShortcutActivity::class.java)
-            intent.putExtra(SHORTCUT, P3)
-            startActivity(intent)
-        }
-        img_btn_deleteP3.setOnClickListener {
-            mapper.remove(Mapper.ProgKey.P3)
-            refresh()
+        var delPos = 1
+        listOf<ImageButton>(img_btn_deleteP1, img_btn_deleteP2, img_btn_deleteP3).forEach {
+            val pNameDel = "P${delPos++}"
+            it.setOnClickListener {
+                when (pNameDel) {
+                    "P1" -> {
+                        mapper.remove(Mapper.ProgKey.P1)
+                        refresh()
+                    }
+                    "P2" -> {
+                        mapper.remove(Mapper.ProgKey.P2)
+                        refresh()
+                    }
+                    "P3" -> {
+                        mapper.remove(Mapper.ProgKey.P3)
+                        refresh()
+                    }
+                }
+            }
         }
     }
